feature: add presentation layer to books module

+++ features/books/src/main/java/com/cgomezq/bookstore/features/books/domain/entities/Book.kt
@@ -1,7 +1,7 @@
 package com.cgomezq.bookstore.features.books.domain.entities
 
 data class Book(
-    val isbn: Int,
+    val isbn: Long,
     val title: String,
     val author: String,
     val summary: String,
